move init functions into fluidkeys/init.go
